+++ src/redteam/strategies/index.ts
@@ -3,7 +3,7 @@ import dedent from 'dedent';
 import logger from '../../logger';
 import type { TestCase, TestCaseWithPlugin } from '../../types';
 import { addBase64Encoding } from './base64';
-import { addCrescendo } from './conversation';
+import { addCrescendo } from './crescendo';
 import { addInjections } from './injections';
 import { addIterativeJailbreaks } from './iterative';
 import { addLeetspeak } from './leetspeak';
